pyproject.toml
[project]
name = "ticket-triage-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.9",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
]
